put manage users button on dashboard page behind feature flag
